+++ src/main/java/ur/edu/pl/project/repositories/ProjectRepository.java
@@ -11,7 +11,7 @@ import ur.edu.pl.project.model.Project;
 import ur.edu.pl.project.model.User;
 
 @RepositoryRestResource(collectionResourceRel = "project", path = "project")
-public interface ProjectRepository extends JpaRepository<Project, Long> {
+public interface ProjectRepository extends JpaRepository<Project, Integer> {
 
     List<Project> findByClient(@Param("client") String client);
